Add CancellationToken support to LanguageClientConnection
This change adds and extra parameter to LanguageClientConnection._sendRequest which accepts the vscode-jsonrpc library's CancellationToken class which can be used to cancel active requests.  When a request is cancelled using one of these tokens, a `$/cancelRequest` notification will be sent to the language server to request a cancellation, but the server must implement cancellation behavior for this to be effective.

This change also adds auto-cancellation of previous requests in the AutocompleteAdapter and OutlineViewAdapter classes.

Related to #140.