Add configuration watching to send `workspace/didChangeConfiguration`
This change adds logic to AutoLanguageClient so that a language package can opt-in to having a specific configuration section watched for changes so that workspace/didChangeConfiguration notifications are sent to the language server.  Language packages can opt-in by implementing the `getRootConfigurationKey` method and optionally the `filterConfigurationObject` method to manipulate the configuration object before it gets sent to the language server.