Gracefully handle unexpected server shutdown
This change adds some logic to gracefully handle sudden server shutdown
like when the language server process gets killed by the user in their
task manager.  The solution is to watch for the RPC connection to get
closed unexpectedly and then dispose of the server in ServerManager
so that no further requests are sent to it.