Fixes subsequent completion request context setting
Currently, after a trigger character, if the user keeps typing, Atom might send
additional completion requests if the initial completion requests did not return
the full set of matches (`isIncomplete = true`). These followup requests are
currently sent with the context `TriggerCharacter`. But they should be sent with
the context `TriggerForIncompleteCompletions`. This commit changes them to be
sent with that.