Fixes textEdit completions
Currently the original replacement prefix sent by the LSP server in the
`textEdit` field is overwritten in getSuggestions(). This, for example,
causes the following case:

    foo->

to autocomplete to

    foo->->bar()

instead of the expected

    foo->bar()

This commit fixes the issue by saving the original replacement prefix
(`->` in the example above), and modifying `getSuggestions()` to append
to instead of overwrite this original replacement prefix.